+++ mne/tests/test_viz.py
@@ -111,7 +111,7 @@ def test_plot_evoked():
 
     # test selective updating of dict keys is working.
     evoked.plot(hline=[1], units=dict(mag='femto foo'))
-    evoked_delayed_ssp.plot(toggle_proj=True)
+    evoked_delayed_ssp.plot(proj='interactive')
     evoked_delayed_ssp.apply_proj()
     assert_raises(RuntimeError, evoked_delayed_ssp.plot, proj='interactive')
     evoked_delayed_ssp.info['projs'] = []